my_todo_app/todo-final/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.0.352
openai==1.3.7
python-dotenv==1.0.0
qdrant-client==1.6.2
langchain-qdrant==0.0.3